refactor: update Ship CoreData model
